+++ core/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0] - 2025-03-18
+
 ### Added
 
 - Added `DeindentPackageDirectives` to match the indentation of compiler directives in dpk files